add module code
add module code

Note: there are a few warning messages,
when I use checkpatch.pl.
Most of them are about printk.
It seems that a few other functions are recommended
before using printk.
I choose to ignore them.
Because we are making a simple module for the learning purpose,
we don't need to worry about other functions.

Signed-off-by: Bin Zeng <zengbin.janis@gmail.com>
Signed-off-by: Joel Savitz <jsavitz@redhat.com>